Change value property type from float to decimal

+++ src/Iot.Domain/Interfaces/IMeasurement.cs
@@ -8,6 +8,6 @@ namespace Iot.Domain.Interfaces
         string Device { get; }
         SensorType SensorType { get; }
         DateTime FullDate { get; }
-        float Value { get; }
+        decimal Value { get; }
     }
 }
